third_party/WebKit/LayoutTests/fast/alignment/parse-align-items.html

<!DOCTYPE html>
<html>
<head>
<style>
#alignItemsBaseline {
    align-items: baseline;
}

#alignItemsLastBaseline {
    align-items: last-baseline;
}

#alignItemsStretch {
    align-items: stretch;
}

#alignItemsStart {
    align-items: start;
}

#alignItemsEnd {
    align-items: end;
}

#alignItemsCenter {
    align-items: center;
}

#alignItemsSelfStart {
    align-items: self-start;
}

#alignItemsSelfEnd {
    align-items: self-end;
}

#alignItemsLeft {
    align-items: left;
}

#alignItemsRight {
    align-items: right;
}

#alignItemsFlexStart {
    align-items: flex-start;
}

#alignItemsFlexEnd {
    align-items: flex-end;
}

#alignItemsEndUnsafe {
    align-items: end unsafe;
}

#alignItemsCenterUnsafe {
    align-items: center unsafe;
}

#alignItemsSelfEndSafe {
    align-items: self-end safe;
}

#alignItemsSelfStartSafe {
    align-items: self-start safe;
}

#alignItemsRightSafe {
    align-items: right safe;
}

#alignItemsLeftUnsafe {
    align-items: left unsafe;
}

#alignItemsFlexStartUnsafe {
    align-items: flex-start unsafe;
}

#alignItemsFlexEndSafe {
    align-items: flex-end safe;
}

</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="alignItemsBaseline"></div>
<div id="alignItemsLastBaseline"></div>
<div id="alignItemsStretch"></div>
<div id="alignItemsStart"></div>
<div id="alignItemsEnd"></div>
<div id="alignItemsCenter"></div>
<div id="alignItemsSelfStart"></div>
<div id="alignItemsSelfEnd"></div>
<div id="alignItemsLeft"></div>
<div id="alignItemsRight"></div>
<div id="alignItemsFlexStart"></div>
<div id="alignItemsFlexEnd"></div>

<div id="alignItemsEndUnsafe"></div>
<div id="alignItemsCenterUnsafe"></div>
<div id="alignItemsSelfEndSafe"></div>
<div id="alignItemsSelfStartSafe"></div>
<div id="alignItemsRightSafe"></div>
<div id="alignItemsLeftUnsafe"></div>
<div id="alignItemsFlexStartUnsafe"></div>
<div id="alignItemsFlexEndSafe"></div>
<script src="resources/alignment-parsing-utils.js"></script>
<script>
description('Test that setting and getting align-items works as expected');

debug("Test getting align-items set through CSS");
var alignItemsBaseline = document.getElementById("alignItemsBaseline");
shouldBe("getComputedStyle(alignItemsBaseline, '').getPropertyValue('align-items')", "'baseline'");

var alignItemsLastBaseline = document.getElementById("alignItemsLastBaseline");
shouldBe("getComputedStyle(alignItemsLastBaseline, '').getPropertyValue('align-items')", "'last-baseline'");

var alignItemsStretch = document.getElementById("alignItemsStretch");
shouldBe("getComputedStyle(alignItemsStretch, '').getPropertyValue('align-items')", "'stretch'");

var alignItemsStart = document.getElementById("alignItemsStart");
shouldBe("getComputedStyle(alignItemsStart, '').getPropertyValue('align-items')", "'start'");

var alignItemsEnd = document.getElementById("alignItemsEnd");
shouldBe("getComputedStyle(alignItemsEnd, '').getPropertyValue('align-items')", "'end'");

var alignItemsCenter = document.getElementById("alignItemsCenter");
shouldBe("getComputedStyle(alignItemsCenter, '').getPropertyValue('align-items')", "'center'");

var alignItemsSelfEnd = document.getElementById("alignItemsSelfEnd");
shouldBe("getComputedStyle(alignItemsSelfEnd, '').getPropertyValue('align-items')", "'self-end'");

var alignItemsSelfStart = document.getElementById("alignItemsSelfStart");
shouldBe("getComputedStyle(alignItemsSelfStart, '').getPropertyValue('align-items')", "'self-start'");

var alignItemsLeft = document.getElementById("alignItemsLeft");
shouldBe("getComputedStyle(alignItemsLeft, '').getPropertyValue('align-items')", "'left'");

var alignItemsRight = document.getElementById("alignItemsRight");
shouldBe("getComputedStyle(alignItemsRight, '').getPropertyValue('align-items')", "'right'");

var alignItemsFlexStart = document.getElementById("alignItemsFlexStart");
shouldBe("getComputedStyle(alignItemsFlexStart, '').getPropertyValue('align-items')", "'flex-start'");

var alignItemsFlexEnd = document.getElementById("alignItemsFlexEnd");
shouldBe("getComputedStyle(alignItemsFlexEnd, '').getPropertyValue('align-items')", "'flex-end'");

var alignItemsEndUnsafe = document.getElementById("alignItemsEndUnsafe");
shouldBe("getComputedStyle(alignItemsEndUnsafe, '').getPropertyValue('align-items')", "'end unsafe'");

var alignItemsCenterUnsafe = document.getElementById("alignItemsCenterUnsafe");
shouldBe("getComputedStyle(alignItemsCenterUnsafe, '').getPropertyValue('align-items')", "'center unsafe'");

var alignItemsSelfEndSafe = document.getElementById("alignItemsSelfEndSafe");
shouldBe("getComputedStyle(alignItemsSelfEndSafe, '').getPropertyValue('align-items')", "'self-end safe'");

var alignItemsSelfStartSafe = document.getElementById("alignItemsSelfStartSafe");
shouldBe("getComputedStyle(alignItemsSelfStartSafe, '').getPropertyValue('align-items')", "'self-start safe'");

var alignItemsRightSafe = document.getElementById("alignItemsRightSafe");
shouldBe("getComputedStyle(alignItemsRightSafe, '').getPropertyValue('align-items')", "'right safe'");

var alignItemsLeftUnsafe = document.getElementById("alignItemsLeftUnsafe");
shouldBe("getComputedStyle(alignItemsLeftUnsafe, '').getPropertyValue('align-items')", "'left unsafe'");

var alignItemsFlexStartUnsafe = document.getElementById("alignItemsFlexStartUnsafe");
shouldBe("getComputedStyle(alignItemsFlexStartUnsafe, '').getPropertyValue('align-items')", "'flex-start unsafe'");

var alignItemsFlexEndSafe = document.getElementById("alignItemsFlexEndSafe");
shouldBe("getComputedStyle(alignItemsFlexEndSafe, '').getPropertyValue('align-items')", "'flex-end safe'");

debug("");
debug("Test initial value of align-items through JS");
element = document.createElement("div");
document.body.appendChild(element);
shouldBe("getComputedStyle(element, '').getPropertyValue('align-items')", "'normal'");

debug("");
debug("Test getting and setting align-items through JS");
element = document.createElement("div");
document.body.appendChild(element);
element.style.alignItems = "center";
checkValues(element, "alignItems", "align-items",  "center", "center");

element.style.alignItems = "unsafe start";
checkValues(element, "alignItems", "align-items",  "start unsafe", "start unsafe");

element.style.alignItems = "flex-end safe";
checkValues(element, "alignItems", "align-items",  "flex-end safe", "flex-end safe");

element.style.alignItems = "right";
checkValues(element, "alignItems", "align-items",  "right", "right");

element.style.alignItems = "center";
checkValues(element, "alignItems", "align-items",  "center", "center");

element.style.alignItems = "self-start";
checkValues(element, "alignItems", "align-items",  "self-start", "self-start");

// The 'auto' value is not valid for the align-items property.
element.style.alignItems = "";
element.style.alignItems = "auto";
checkValues(element, "alignItems", "align-items",  "", "normal");

element.style.display = "flex";
element.style.alignItems = "auto";
checkValues(element, "alignItems", "align-items",  "", "normal");

element.style.display = "grid";
element.style.alignItems = "auto";
checkValues(element, "alignItems", "align-items",  "", "normal");

element.style.alignItems = "self-end";
checkValues(element, "alignItems", "align-items",  "self-end", "self-end");

debug("");
debug("Test bad combinations of align-items");
element = document.createElement("div");
document.body.appendChild(element);

checkBadValues(element, "alignItems", "align-items",  "auto");
checkBadValues(element, "alignItems", "align-items",  "unsafe auto");
checkBadValues(element, "alignItems", "align-items",  "auto safe");
checkBadValues(element, "alignItems", "align-items",  "auto left");
checkBadValues(element, "alignItems", "align-items",  "baseline safe");
checkBadValues(element, "alignItems", "align-items",  "baseline center");
checkBadValues(element, "alignItems", "align-items",  "stretch unsafe");
checkBadValues(element, "alignItems", "align-items",  "stretch right");
checkBadValues(element, "alignItems", "align-items",  "unsafe unsafe");
checkBadValues(element, "alignItems", "align-items",  "unsafe safe");
checkBadValues(element, "alignItems", "align-items",  "center start");
checkBadValues(element, "alignItems", "align-items",  "stretch unsafe");
checkBadValues(element, "alignItems", "align-items",  "safe stretch");
checkBadValues(element, "alignItems", "align-items",  "baseline safe");
checkBadValues(element, "alignItems", "align-items",  "unsafe baseline");
checkBadValues(element, "alignItems", "align-items",  "unsafe safe left");
checkBadValues(element, "alignItems", "align-items",  "unsafe left safe");
checkBadValues(element, "alignItems", "align-items",  "left safe unsafe safe");
checkBadValues(element, "alignItems", "align-items",  "legacy start");
checkBadValues(element, "alignItems", "align-items",  "legacy end");
checkBadValues(element, "alignItems", "align-items",  "legacy right unsafe");
checkBadValues(element, "alignItems", "align-items",  "legacy auto");
checkBadValues(element, "alignItems", "align-items",  "legacy stretch");
checkBadValues(element, "alignItems", "align-items",  "legacy");
checkBadValues(element, "alignItems", "align-items",  "legacy left right");

debug("");
debug("Test the value 'initial'");
element.style.display = "";
checkInitialValues(element, "alignItems", "align-items", "center", "normal");

debug("");
debug("Test the value 'initial' for grid containers");
element.style.display = "grid";
checkInitialValues(element, "alignItems", "align-items", "left safe", "normal");

debug("");
debug("Test the value 'initial' for flex containers");
element.style.display = "flex";
checkInitialValues(element, "alignItems", "align-items", "right unsafe", "normal");

debug("");
debug("Test the value 'inherit'");
checkInheritValues("alignItems", "align-items", "end");
checkInheritValues("alignItems", "align-items", "left safe");
checkInheritValues("alignItems", "align-items", "center unsafe");

</script>
</body>
</html>
